requirements.txt
PyMySQL==1.1.0
PyYAML==6.0.1
sqlalchemy[asyncio]==2.0.20
alembic==1.11.3
elasticsearch[async]==8.9.0
passlib==1.7.4
sentry-sdk==1.29.2
redis==5.0.0
requests==2.31.0
click==8.1.7
honcho==1.1.0
aiosmtplib==2.0.2
pydantic==2.2.1
pydantic-settings==2.0.3
aiomysql==0.2.0
pytest==7.4.0
orjson==3.9.5
python-dateutil
aiofile==3.8.8
fastapi==0.101.1
uvicorn==0.23.2
jinja2==3.1.2
arq==0.25.0
pytest_asyncio==0.21.1
httpx==0.24.1
python-multipart==0.0.6
email-validator==2.0.0.post2
freezegun==1.2.2
uuid6==2023.5.2
aiorun==2023.7.2
aiosqlite==0.19.0
pyjwt==2.8.0
respx==0.20.2